requirements.txt
Flask
reportlab
supabase
python-dotenv
gunicorn
